feat: add environment model utility methods
