+++ src/components/Models/EnvironmentModel/EnvironmentModel.container.tsx
@@ -1,5 +1,6 @@
 import React, { FC, Suspense, useState } from 'react';
 import { EnvironmentModel, EnvironmentModelProps } from './EnvironmentModel.component';
+import { EnvironmentModels, environmentModels } from 'src/services/Environment.service';
 
 /**
  * Contains model to handle suspense fallback.
@@ -7,10 +8,13 @@ import { EnvironmentModel, EnvironmentModelProps } from './EnvironmentModel.comp
 export const EnvironmentModelContainer: FC<EnvironmentModelProps> = (props) => {
   /* eslint-disable-next-line react/jsx-no-useless-fragment */
   const [fallback, setFallback] = useState<JSX.Element>(<></>);
+  const { environment } = props;
+  const isStaticPreset = environment in environmentModels;
+  const environmentSrc = isStaticPreset ? environmentModels[environment as EnvironmentModels] : environment;
 
   return (
     <Suspense fallback={fallback}>
-      <EnvironmentModel setModelFallback={setFallback} {...props} />
+      <EnvironmentModel setModelFallback={setFallback} {...props} environment={environmentSrc} />
     </Suspense>
   );
 };
